Finalised initial 4 layer setup of ferris
Yet to add media controls or brightness etc, and there are loads of keys
and other functionality that I would still like to incorporate into the
layout (plus this isn't final and the layout subject to heavy changes),
but it feels good to finally be able to use the keyboard and it not miss
keys I need.

The alpha layout is Colemak-DH.

Commented out code is related to a custom "repeat" key that can repeat
the last key combination. This seems like it would be a very powerful
feature to include, like vim's . command, but it seemed to introduce
some input lag that I wanted to eliminate before I finished the
provisional layout. I'm going to try introducing it back in shortly,
after I've used the keyboard for a bit first.